Tree system refactoring.
