+++ src/main/java/rtg/world/gen/feature/tree/rtg/pinaceae/TreeRTGPinaceaeBase.java
@@ -1,4 +1,4 @@
-package rtg.world.gen.feature.tree.rtg;
+package rtg.world.gen.feature.tree.rtg.pinaceae;
 
 import java.util.Random;
 
@@ -7,10 +7,10 @@ import net.minecraft.block.material.Material;
 import net.minecraft.world.World;
 import rtg.world.gen.feature.tree.TreeRTG;
 
-public class TreeRTGPinaceae extends TreeRTG
+public class TreeRTGPinaceaeBase extends TreeRTG
 {
 
-	public TreeRTGPinaceae()
+	public TreeRTGPinaceaeBase()
 	{
 		super();
 	}
