+++ src/main/java/rtg/world/gen/feature/tree/rtg/pinaceae/picea/TreeRTGPinaceaePiceaBase.java
@@ -1,16 +1,16 @@
-package rtg.world.gen.feature.tree.rtg.pinaceae;
+package rtg.world.gen.feature.tree.rtg.pinaceae.picea;
 
 import java.util.Random;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.material.Material;
 import net.minecraft.world.World;
-import rtg.world.gen.feature.tree.rtg.TreeRTGPinaceae;
+import rtg.world.gen.feature.tree.rtg.pinaceae.TreeRTGPinaceaeBase;
 
-public class TreeRTGPinaceaePicea extends TreeRTGPinaceae
+public class TreeRTGPinaceaePiceaBase extends TreeRTGPinaceaeBase
 {
 
-	public TreeRTGPinaceaePicea()
+	public TreeRTGPinaceaePiceaBase()
 	{
 		super();
 	}
