+++ src/main/java/rtg/world/gen/feature/tree/rtg/pinaceae/picea/TreeRTGPinaceaePiceaSitchensis.java
@@ -5,9 +5,8 @@ import java.util.Random;
 import net.minecraft.block.Block;
 import net.minecraft.init.Blocks;
 import net.minecraft.world.World;
-import rtg.world.gen.feature.tree.rtg.pinaceae.TreeRTGPinaceaePicea;
 
-public class TreeRTGPinaceaePiceaSitchensis extends TreeRTGPinaceaePicea
+public class TreeRTGPinaceaePiceaSitchensis extends TreeRTGPinaceaePiceaBase
 {
 
 	public TreeRTGPinaceaePiceaSitchensis()
