+++ src/main/java/rtg/world/gen/feature/tree/rtg/pinaceae/pinus/TreeRTGPinaceaePinusPonderosa.java
@@ -5,9 +5,8 @@ import java.util.Random;
 import net.minecraft.block.Block;
 import net.minecraft.init.Blocks;
 import net.minecraft.world.World;
-import rtg.world.gen.feature.tree.rtg.pinaceae.TreeRTGPinaceaePinus;
 
-public class TreeRTGPinaceaePinusPonderosa extends TreeRTGPinaceaePinus
+public class TreeRTGPinaceaePinusPonderosa extends TreeRTGPinaceaePinusBase
 {
 
 	public TreeRTGPinaceaePinusPonderosa()
